modified batchload api

+++ server/src/main/java/com/bassemHalim/Activity/ActivityController.java
@@ -17,7 +17,7 @@ public class ActivityController {
     @GetMapping("/activity-list")
     ResponseEntity<List<ActivityDTO>> getActivityList(
             @RequestParam(required = false, defaultValue = "20") int limit,
-            @RequestParam(required = false, defaultValue = "0") int start,
+            @RequestParam(required = false, defaultValue = "0") long start,
             @RequestParam(required = false, defaultValue = "false") boolean GarminSync
     ) {
         List<ActivityDTO> activityList = activityService.getActivityList(start, limit, GarminSync);
